Update leads  djc logic

+++ app/Repositories/Focus/lead/LeadRepository.php
@@ -74,7 +74,7 @@ class LeadRepository extends BaseRepository
      */
     public function delete(Lead $lead)
     {
-        if (count($lead->quotes)) return;          
+        if (count($lead->quotes) || count($lead->djcs)) return;          
         else return $lead->delete();
         
         throw new GeneralException(trans('exceptions.backend.productcategories.delete_error'));
